chanfe byte to int , reason to catch negative value in validation

+++ ExamApp.Services/DTOs/ExamDto.cs
@@ -5,6 +5,6 @@
         public required string SubjectCode { get; init; } = string.Empty;
         public required int StudentNumber { get; init; }
         public required string Date { get; init; }      // date as string иначе придется лепить правила для конвертации и мне лениво
-        public required byte Mark { get; init; }
+        public required int Mark { get; init; }
     }
 }
